+++ shell/commands/makeself/hello.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+makeself --gzip --sha256 hello hello_installer.run "Hello" ./hello/installer.sh